emotes in title performance improvements, added inside clips
